reworked placeShip() function to have reverse coordinates
I am passing the coordinates in reverse order like
this: [yPosition first (letter), xPosition second (number)]
the reason I am doing it this way is because usually I call
out the letter, then the number. I also number type converted
to ensure that the numbers aren't accidentally strings.
I need to work on a check during ship placement that checks if
the ship's future coordinates in the direction it is facing
interferes with any existing placed ships on the board already.