Fixed loop bug with checkDirectionInBounds() and checkForShipClash()
The checkDirectionInBounds() function was flagging
valid ships incorrectly because the for loop
would iterate an extra time, so I added 1 to the
iterator (i) to prevent it from over-iterating and
accidentally triggering the if statment. Also
fixed that with the for loop in checkForShipClash()
and the for loop in placeShip() for consistency
between for loops.