test: added a receiveAttack() test to test for successful attack
The idea is when the gameboard receives an attack, it will
return true if the attack hits or false if the attack misses.
If the attack hits, then it will record the coordinates to the attackHistory Map()
object, and store the ship object as the value. If the attack misses,
it still stores the coordinates as the key but the value will be
a string "miss" to indicate a miss. This is so that I can
track what spaces have been shot at already and what spaces
have been hit or missed.