feat: finished receiveAttack() function
Tested that the receiveAttack() works as intended.
Returns true if the hit is successful, increments the
ship at the found location by calling it's hit() function
and storing the coordinate and ship into attackHistory.
Returns false if the coordinates are out of bounds,
if the coordinates given has already been attacked already, and
if the attack is a miss. If it is a miss, it will store
the coordinates in the attackHistory and the string "miss".