addon: added a ship() isSunk() check when the attack is a successful hit
Whenever an attack is received that is successful in hitting the
ship, it will take the ship that got hit, call it's hit() function
and then call the ship's isSunk() function to also check if that
hit caused the ship to sink based on it's total hits and the
ship() object's length. This is also to help check
if the entire board is won or not based on the sunk condition
on every ship.