feat: checkAllSunk() function is finished
The gameboard uses a gamePieces Map() with references to
the shipObj on the board. checkAllSunk() checks
each ship's sunk status, and if there is at least 1
ship that ISN'T sunk, then checkAllSunk returns false
to indicate there is still at least 1 living ship.
If it doesn't detect any alive ships, it will return
true, indicating that all ships are sunk on the board.