working: makeBoard() function player input works
The input parameters of the makeBoard() function
allows the passing of all ship coordinates, and
orientations to the function. If the user is
a real player, the function utilizes the
.placeShip() function from the gameboard() to
place the ships with the input parameters. I have
confirmed this works and inspected the board
to have at least 17 entries (all ship lengths
added together you should have a total of 17
spaces that all ships are taking up on the board).
I need to work on the computer randomly generated portion
of the makeBoard function.