feat: finished makeBoard() computer section and utility functions
finished making the computer part of the makeBoard()
function. If the player is a computer, the code will generate
random coordinate locations and a random direction, then it
will use a loop to keep looping and generating directions until
the ship placement is valid (returns true), then it will stop the loop
and move on. I also created utility functions to check if the
player lost, receive attacks, and sending attacks.