working: makePlayer1() and makePlayer2() also utility functions
made a makePlayer1 and makePlayer2 function that initializes
the two main players in the game. Makes both with
ship placements, and player 2 has the option of being
a computer player which will place ships in random locations.
These two methods fully setup the game with both players
and ship positions.