working: rewriting gamecontroller/module to now drive game using DOM
I was confused on how to run my game, but when I read back
on the odin project it says that after the gameboard object
is made, it is now appropriate to start making the UI to
drive the game using event listeners with the objects I made.
I am going to rework the gamecontroller module to interact with
the DOM elements and using the objects I made.