feat: made a playRound() function in gamecontroller
playRound() plays an entire round of battleship
depending on the currentTurn. If its player 1's turn, then
it will take a given attack coordinate, translate it to
an array, then send the attack to player 2's board where player 2
board will receive the attack, then it will update the attack history
on player 2's board, then check for win condition. If it is not met,
then swap turns. Same idea with player 2's turn. If the player 2
is a computer, then it will get the randomly generated
attack coordinate from the sendAttack() player method, then
send it to player 1's board, then check for win condition and swap
turns if the win condition is not met. I still need to test
if the win condition activates properly when all ships are sunk.