working: working on playRound() to get win condition
playRound() is a function with two parameters,
the Y coordinate (letter) and the X coordinate (number).
The playRound() will behave differently depending
on whose turn it is, and specifically when the player 2
is a computer, it will generate a random valid
attack coordinate depedning on if its within bounds
and hasn't been attacked before. After making an
attack, it checks to see if the game is won by either
side by checking if the opposing player who is being
attacked has all their ships sunk. If not, the game
continues and the current turn swaps between player 1
and player 2. If it does meet a successful win
condition, then it will return a boolean true
signaling the end of the game.