working: interactions between modals
So the sequence of events I coded are:
1. game mode selection
if pvp:
    - go to player 1 option and create player 1
    - go to player 2 option and create player 2
    - start game
if pvc:
    - go to player 1 option and create player 1,
    then also create a computer player 2
    - start game

I was struggling with the event handlers for
each case, since in both game modes, player 1
options are displayed, so I needed to
add a playerVsComputer boolean into renderPlayer1Options()
to ensure that depending on which game mode is selected,
the player 1 options will transition into the correct order
as listed above. At the end of the day, it transitions
and creates the player entries properly. I need to
add form validation later, but for now the sequence works.
I also added the "name" attributes to all the
player inputs to give the values names when parsing them for
form data.